smallf1: factor naze32/cc/cc3d common code
Also changes lots of COPTERCONTROL defines to SMALLF1 (but COPTERCONTROL
is still used in Modules/Attitude/smallf1).
