+++ flight/Modules/ManualControl/geofence_control.c
@@ -103,7 +103,7 @@ int32_t geofence_control_select(bool reset_controller)
 //! Query if out of bounds and the geofence controller should take over
 bool geofence_control_activate()
 {
-#if (!defined(COPTERCONTROL) && !defined(GIMBAL))
+#if (!defined(SMALLF1) && !defined(GIMBAL))
 	// Check if the module is running
 	if (GeoFenceSettingsHandle()) {
 		uint8_t alarm_status[SYSTEMALARMS_ALARM_NUMELEM];
